requirements.txt
aiogram==2.11.2
requests==2.7.0
python-dotenv==0.15.0
pandas==1.1.3
numpy==1.19.2
matplotlib==3.2